backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "elevenlabs>=2.13.0",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.32.0",
    "openai>=1.102.0",
    "pillow>=11.3.0",
]
